+++ Starwar/Assets/Scripts/Player Control/AI/SteeringAgent.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using UnityEngine;
-using BehaviorTree;
 public class SteeringAgent : MonoBehaviour
 {
 
@@ -75,109 +74,3 @@ public class SteeringAgent : MonoBehaviour
     }
 
 }
-
-public class IsAssaultState : BehaviorNode
-{
-    public FighterBT tree;
-    public IsAssaultState(FighterBT tree)
-    {
-        this.tree = tree;
-    }
-
-    public override BehaviorNodeState Evaluate()
-    {
-        if (tree.state == FighterBT.State.Assault)
-        { return BehaviorNodeState.SUCCESS; }
-        else
-        { return BehaviorNodeState.FAILURE; }
-    }
-}
-
-public class SetStateAssault : BehaviorNode
-{
-    public FighterBT tree;
-    public SteeringAgent agent;
-    public SetStateAssault(FighterBT tree, SteeringAgent steeringAgent)
-    {
-        this.tree = tree;
-        agent = steeringAgent;
-    }
-
-    public override BehaviorNodeState Evaluate()
-    {
-        tree.state = FighterBT.State.Assault;
-        agent.Assault();
-        return BehaviorNodeState.SUCCESS;
-    }
-}
-
-public class IsPatrolState : BehaviorNode
-{
-    public FighterBT tree;
-    public IsPatrolState(FighterBT tree)
-    {
-        this.tree = tree;
-    }
-
-    public override BehaviorNodeState Evaluate()
-    {
-        if (tree.state == FighterBT.State.Patrol)
-        { return BehaviorNodeState.SUCCESS; }
-        else
-        { return BehaviorNodeState.FAILURE; }
-    }
-}
-
-public class SetStatePatrol : BehaviorNode
-{
-    public FighterBT tree;
-    public SteeringAgent agent;
-    public SetStatePatrol(FighterBT tree, SteeringAgent steeringAgent)
-    {
-        this.tree = tree;
-        agent = steeringAgent;
-    }
-
-    public override BehaviorNodeState Evaluate()
-    {
-        tree.state = FighterBT.State.Patrol;
-        agent.Patrol();
-        return BehaviorNodeState.SUCCESS;
-    }
-}
-
-public class IsDeadState : BehaviorNode
-{
-    public FighterBT tree;
-    public IsDeadState(FighterBT tree)
-    {
-        this.tree = tree;
-    }
-
-    public override BehaviorNodeState Evaluate()
-    {
-        if (tree.state == FighterBT.State.Dead)
-        { return BehaviorNodeState.SUCCESS; }
-        else
-        { return BehaviorNodeState.FAILURE; }
-    }
-}
-
-public class SetStateDead : BehaviorNode
-{
-    public FighterBT tree;
-    public SteeringAgent agent;
-    public SetStateDead(FighterBT tree, SteeringAgent steeringAgent)
-    {
-        this.tree = tree;
-        agent = steeringAgent;
-    }
-
-    public override BehaviorNodeState Evaluate()
-    {
-        tree.state = FighterBT.State.Dead;
-        agent.Dead();
-        return BehaviorNodeState.SUCCESS;
-    }
-}
-
