+++ Starwar/Assets/Scripts/Player Control/Health.cs
@@ -1,5 +1,4 @@
-﻿using BehaviorTree;
-using UnityEngine;
+﻿using UnityEngine;
 public class Health : MonoBehaviour
 {
     public float MaxHealth;
@@ -22,7 +21,6 @@ public class Health : MonoBehaviour
             else { currentHealth = value; }
         }
     }
-
     public bool IsAlive { get { return isAlive; } }
 
     private void Start()
@@ -39,43 +37,5 @@ public class Health : MonoBehaviour
             CurrentHealth -= damage;
         }
     }
-
 }
-
-public class AmIHealthy : BehaviorNode
-{
-    public Health myHealth;
-    public float criticalPercentage;
-    public AmIHealthy(Health health, float criticalPercentage)
-    {
-        myHealth = health;
-        this.criticalPercentage = criticalPercentage;
-    }
-
-    public override BehaviorNodeState Evaluate()
-    {
-        float percentage = myHealth.CurrentHealth / myHealth.MaxHealth;
-        if (percentage < criticalPercentage)
-        {
-            return BehaviorNodeState.FAILURE;
-        }
-        else
-        {
-            return BehaviorNodeState.SUCCESS;
-        }
-    }
-}
-
-public class AmIAlive : BehaviorNode
-{
-    public Health myHealth;
-    public AmIAlive(Health health)
-    { myHealth = health; }
-
-    public override BehaviorNodeState Evaluate()
-    {
-        return myHealth.IsAlive ? BehaviorNodeState.SUCCESS : BehaviorNodeState.FAILURE;
-    }
-}
-
