+++ Starwar/Assets/Scripts/Player Control/MachineGun.cs
@@ -28,6 +28,4 @@ public class MachineGun : MonoBehaviour
 
         lastFire = Time.fixedTime;
     }
-
-
 }
