--- Starwar/Assets/Scripts/Player Control/Test.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a71d61bf31eae304597d8c79be976527
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
